all: Move pkg/migrate into pkg/postgres
