+++ controller/schema.go
@@ -2,11 +2,11 @@ package main
 
 import (
 	"github.com/flynn/flynn/Godeps/_workspace/src/github.com/flynn/go-sql"
-	"github.com/flynn/flynn/pkg/migrate"
+	"github.com/flynn/flynn/pkg/postgres"
 )
 
 func migrateDB(db *sql.DB) error {
-	m := migrate.NewMigrations()
+	m := postgres.NewMigrations()
 	m.Add(1,
 		`CREATE EXTENSION IF NOT EXISTS "uuid-ossp"`,
 		`CREATE EXTENSION IF NOT EXISTS "hstore"`,
