+++ pkg/postgres/migrate.go
@@ -1,4 +1,4 @@
-package migrate
+package postgres
 
 import "github.com/flynn/flynn/Godeps/_workspace/src/github.com/flynn/go-sql"
